Add ohos_app gn template
Description: add ohos_app  gn template and build script

Issue:I6SHSR

Test:build

+++ ohos/packages/modules_install.py
@@ -107,11 +107,19 @@ def copy_modules(system_install_info, install_modules_info_file,
                 dest = dest[1:]
             dest_list.append(dest)
             # dest_dir_prefix
-            dest_dir = os.path.join(platform_installed_path,
-                                    os.path.dirname(dest))
+            if os.path.isfile(source):
+                dest_dir = os.path.join(platform_installed_path,
+                                        os.path.dirname(dest))
+            elif os.path.isdir(source):
+                dest_dir = os.path.join(platform_installed_path, dest)
             if not os.path.exists(dest_dir):
                 os.makedirs(dest_dir)
-            shutil.copy(source, os.path.join(platform_installed_path, dest))
+            if os.path.isfile(source):
+                shutil.copy(source, os.path.join(platform_installed_path, dest))
+            elif os.path.isdir(source):
+                for file_name in os.listdir(source):
+                    file_path = os.path.join(source, file_name)
+                    shutil.copy(file_path, os.path.join(platform_installed_path, dest, file_name))
 
         # add symlink
         if 'symlink' in module_info:
